requirements.txt
asgiref==3.8.1
Django==5.2.1
lxml==5.4.0
pillow==11.2.1
python-dotenv==1.1.0
python-pptx==1.0.2
sqlparse==0.5.3
typing_extensions==4.13.2
XlsxWriter==3.2.3
